# Make leaf insertion and hashing recursive