--- a/akd/Cargo.toml
+++ b/akd/Cargo.toml
@@ -39,7 +39,6 @@ curve25519-dalek = "3"
 dashmap = { version = "5" }
 ed25519-dalek = "1"
 hex = "0.4"
-keyed_priority_queue = "0.3"
 log = { version = "0.4.8", features = ["kv_unstable"] }
 tokio = { version = "1.10", features = ["sync", "time", "rt"] }