# Overview
To parallelize leaf insertion and hashing, the process needs to be made fully recursive, such that work can be split into subtasks and performed independently of one another. As an additional plus point, a fully recursive approach is more concise and easier to reason about. 

To this end, this PR does the necessary refactor. I do this by partitioning the set of leaves to be inserted based on the current node, then delegating the partitions to the left and right child nodes to handle recursively. When both child nodes return, the current node is hashed and returned to its parent.

The actual parallelization comes in #302, for ease of review.

# Benchmark

**Main**
Cherry-picked the updated benchmark (6c9f83e, 3ade785) on top of main (see: https://github.com/afterdusk/akd/tree/benchmark) to compare this PR against.

<img width="486" alt="image" src="https://user-images.githubusercontent.com/20502803/209638545-6c8178ed-e9fb-43f7-ad54-cbf8f60884fb.png">


**This PR**
<img width="480" alt="image" src="https://user-images.githubusercontent.com/20502803/209639737-48453bd4-0d7e-48cd-b879-b496365b03d5.png">


# Progress
- [x] Rewrite leaf insertion and hashing to be recursive
- [x] Rebase on main branch
- [x] Consider porting some tests from `tree_node.rs` to `append_only_zks.rs`, as tree traversal logic was largely abstracted out of `tree_node.rs` in this PR (will relook in a separate PR)
- [x] Benchmark