@afterdusk 2022-12-27
Rebased on main and addressed review comments. Will work on benchmarking this PR separately from #302, as #302 is currently blocked on some perf issues with parallelization. If this PR does not impact perf negatively, I would like to merge it first.