@kevinlewi 2022-12-27
Really loving how clean this is!

Two places where we might be able to improve on perf:
1. When you are getting the longest prefix of all nodes in the insertion set, instead of iterating through the insertion set and calling longest_common_prefix on each element, we might be able to make use of something like this function (https://github.com/novifinancial/akd/blob/main/akd/src/append_only_zks.rs#L165-L175). Basically, start by assuming that the insertion set is a sorted vector of NodeLabels, sorted by their label. Then, start with a candidate prefix derived from the first element of the sorted vec, and check if it is a prefix of all of the others via binary search. Keep going until you find the point at which you have the "longest" prefix (also could be found via binary search on the bits of the prefix).
2. Kindof related to #1, but right now when you compute `partition_longest_common_prefix`, you are iterating over the insertion set to compute the partition. However, if the insertion set is already sorted, you would just need to find the "midpoint" (again via binary search), and then break up the vec into two partitions immediately.

Hope that makes sense! But anyways let's land this first and worry about these perf optimizations later. The structure of this is really nice.