@afterdusk 2022-12-27
@kevinlewi Thanks for the suggestions, they make sense to me! I'll work on them and should have a PR up by this week :)