@kevinlewi 2022-12-27
@afterdusk Btw I did a checkout of this PR and ran it locally just to see what the perf change is right now,

Main branch (without this PR):
```
Batch insertion (100000 initial leaves) took 9666 ms
Batch insertion (100000 initial leaves, 100000 inserted leaves) took 11955 ms
```
This branch (with this PR):
```
Batch insertion (100000 initial leaves) took 7552 ms
Batch insertion (100000 initial leaves, 100000 inserted leaves) took 8571 ms
```

Already a 20-30% speedup, it seems...!